# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Notes API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/notebooks/{notebook-id}/notes/": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Getting a list of user notes in a notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.GetAllNotesResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Creating a note in a notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Note data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateNoteDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note ID",
                        "schema": {
                            "$ref": "#/definitions/docs.CreateNoteResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
        "/api/notebooks/{notebook-id}/notes/{note-id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Getting a note in a notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "note-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.GetNoteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Deleting a note from a notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "note-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK status",
                        "schema": {
                            "$ref": "#/definitions/docs.OkStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Updating a note in a notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "note-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New note data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PatchNoteDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK status",
                        "schema": {
                            "$ref": "#/definitions/docs.OkStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/api/shared-notes/": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shared note"
                ],
                "summary": "Getting a list of shared notes",
                "responses": {
                    "200": {
                        "description": "Shared notes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.GetSharedNotesInfoResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
        "/api/shared-notes/incoming": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shared note"
                ],
                "summary": "Creating a shared note",
                "parameters": [
                    {
                        "description": "Shared note data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateSharedNoteDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shared note ID",
                        "schema": {
                            "$ref": "#/definitions/docs.CreateSharedNoteResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
        "/api/shared-notes/incoming/{shared-note-id}": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shared note"
                ],
                "summary": "Accepting a shared note",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shared note ID",
                        "name": "shared-note-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK status",
                        "schema": {
                            "$ref": "#/definitions/docs.OkStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shared note"
                ],
                "summary": "Deleting a shared note",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shared note ID",
                        "name": "shared-note-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK status",
                        "schema": {
                            "$ref": "#/definitions/docs.OkStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
        "/api/shared-notes/outgoing/{note-id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shared note"
                ],
                "summary": "Getting an outgoing shared notes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "note-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Outgoing shared notes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.GetOutgoingSharedNotesInfoResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
        "/api/shared-notes/{shared-note-id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shared note"
                ],
                "summary": "Getting a data of shared note",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shared note ID",
                        "name": "shared-note-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data of shared note",
                        "schema": {
                            "$ref": "#/definitions/domain.SharedNoteData"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docs.CreateNoteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "docs.CreateSharedNoteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "docs.GetAllNotesResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "body": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "id": {
                                "type": "integer"
                            },
                            "title": {
                                "type": "string"
                            },
                            "updated_at": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "docs.GetNoteResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "docs.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "docs.OkStatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "domain.CreateNoteDTO": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.CreateSharedNoteDTO": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "note_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.GetOutgoingSharedNotesInfoResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "shared_notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.OutgoingSharedNoteInfo"
                    }
                }
            }
        },
...
        "domain.GetSharedNotesInfoResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "shared_notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SharedNoteInfo"
                    }
                }
            }
        },
...
        "domain.OutgoingSharedNoteInfo": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "recipient_login": {
                    "type": "string"
                },
                "recipient_name": {
                    "type": "string"
                }
            }
        },
...
        "domain.PatchNoteDTO": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "domain.SharedNoteData": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.SharedNoteInfo": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "owner_login": {
                    "type": "string"
                },
                "owner_name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}